+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Fixed
 - php 8.5 compatibility fix
 - Nullpointer in DataloadController::importFile() #560 #555
+- Extend filter variable replacement tests with multiple filter cases
 
 
 ## 6.0.2 - 2025-09-19